Move ImportBeaconKeysFromConfig to beacon.cpp.
